add tickets screen

+++ app/src/main/java/com/tangonoches/teacher/data/responses/ticketEventTypes/TicketEventTypeDto.kt
@@ -3,11 +3,11 @@ package com.tangonoches.teacher.data.responses.ticketEventTypes
 import com.tangonoches.teacher.data.models.TicketEventTypeModel
 
 data class TicketEventTypeDto(
-    val id: Int = 0,
+    val id: Long = 0,
     val name: String = ""
 )
 
-fun TicketEventTypeDto.toModel():TicketEventTypeModel =
+fun TicketEventTypeDto.toModel(): TicketEventTypeModel =
     TicketEventTypeModel(
         id = id,
         name = name
